Onion Browser: link proving Tor Project endorsement (#2179)
In the Onion Browser's description is written that it is endorsed by the Tor Project, but the Onion Browser's website has no easy to find link to the Tor Project's website to prove it. For readers convenience I have added link to https://support.torproject.org where they can read about the Onion Browser from the Tor Project's perspective.

I know, this is not Wikipedia and we do not have to source everything, however, in the Firefox for iOS, there is link to Disconnect's website, just so users can read about their Tracking protection lists, and being able to quickly verify that the Onion Browser is really officially endorsed by the Tor Project is arguably much more important, therefore the link should be included.

Other, more explicit, endorsement on torproject.org is in a small print under Android download page https://torproject.org/download/index.html#android that says "Are you an iOS user? We encourage you to try Onion Browser.", however this one is much less visible, because of that I'm not using this link.